-implement localizaTion service into the templates -omit translation form from implementation due to refactor that needs to be done
