+++ src/main/java/lule/dictionary/service/localization/LocalizationService.java
@@ -8,4 +8,8 @@ public interface LocalizationService {
     Map<String, String> navbarLocalization(Language language);
     Map<String, String> documentListLocalization(Language language);
     Map<String, String> authLocalization (Language language);
+    Map<String, String> createImportLocalization(Language language);
+    Map<String, String> translationFormLocalization(Language language);
+
+
 }
