+++ src/main/jte/auth/signup.jte
@@ -5,6 +5,7 @@
 @param ServiceResultImp<?> result
 @param BaseAttribute baseAttribute
 @param Map<String, String> navbarLocalization
+@param Map<String, String> authLocalization
 
 
 @template.layout.base(authView = @`
@@ -27,9 +28,9 @@
         }'
         class="md:w-2/5
         w-4/5 flex flex-col bg-secondary p-8 mx-auto gap-8">
-        <span class="text-accent text-2xl font-bold">Sign up</span>
+        <span class="text-accent text-2xl font-bold">${authLocalization.get("sign_up")}</span>
         <div class="flex flex-col">
-            <label class="text-accent" for="login">Username</label>
+            <label class="text-accent" for="login">${authLocalization.get("username")}</label>
             <input
             class="
             outline-none border border-2 border-primary rounded text-neutral text-sm min-h-8
@@ -46,7 +47,7 @@
             @endif
         </div>
         <div class="flex flex-col">
-            <label class="text-accent" for="email">Email</label>
+            <label class="text-accent" for="email">${authLocalization.get("email")}</label>
             <input
             class="
             outline-none border border-2 border-primary rounded text-neutral text-sm min-h-8
@@ -63,7 +64,7 @@
             @endif
         </div>
         <div class="flex flex-col">
-            <label class="text-accent" for="password">Password</label>
+            <label class="text-accent" for="password">${authLocalization.get("password")}</label>
             <input
             class="
             outline-none border border-2 border-primary rounded text-neutral text-sm min-h-8
@@ -80,13 +81,13 @@
             @endif
         </div>
         <div class="flex gap-4 items-center justify-between">
-            <button type="submit" class="text-left px-2 py-1 bg-accent text-primary rounded cursor-pointer">Submit</button>
+            <button type="submit" class="text-left px-2 py-1 bg-accent text-primary rounded cursor-pointer">${authLocalization.get("submit")}</button>
             <a
             class="cursor-pointer underline text-accent font-semibold"
             hx-get="/auth/login"
             hx-target="#main"
             hx-swap="outerHTML"
-            hx-push-url="true">Log in</a>
+            hx-push-url="true">${authLocalization.get("log_in")}</a>
         </div>
     </form>
 `,
