+++ src/main/jte/create-import-form/base-form.jte
@@ -4,7 +4,8 @@
 
 @param ServiceResultImp<?> result = null
 @param BaseAttribute baseAttribute 
-@param Map<String, String> navbarLocalization 
+@param Map<String, String> navbarLocalization
+@param Map<String, String> importFormLocalization
 
 
 @template.layout.base(content = @`
@@ -27,7 +28,7 @@
     hx-push-url="true">
         <input type="hidden" name="${baseAttribute._csrf().getParameterName()}" value="${baseAttribute._csrf().getToken()}">
         <div class="flex flex-col gap-2">
-            <label class="text-accent font-bold text-2xl min-h-8" for="title">Title</label>
+            <label class="text-accent font-bold text-2xl min-h-8" for="title">${importFormLocalization.get("title")}</label>
             <input
             class="
             outline-none border border-2 border-primary rounded text-neutral
@@ -45,7 +46,7 @@
             @endif
         </div>
         <div class="flex flex-col gap-4">
-            <span class="text-accent font-bold text-2xl">Content</span>
+            <span class="text-accent font-bold text-2xl">${importFormLocalization.get("content")}</span>
             <div class="flex justify-center">
                 <button class="
                 bg-primary border border-2 border-r-1 border-primary text-accent px-4 py-1 grow-1 outline-none"
@@ -55,7 +56,7 @@
                 hx-target="#content"
                 hx-swap="outerHTML"
                 hx-push-url="false"
-                autofocus>Import by URL</button>
+                autofocus>${importFormLocalization.get("import_by_url")}</button>
                 <button class="
                 border border-2 border-l-1 border-primary text-accent px-4 py-1 grow-1"
                 type="button"
@@ -63,14 +64,17 @@
                 hx-get="/imports/textarea-form"
                 hx-target="#content"
                 hx-swap="outerHTML"
-                hx-push-url="false">Insert manually</button>
+                hx-push-url="false">${importFormLocalization.get("insert_manually")}</button>
             </div>
-            @template.create-import-form.url-form(result = result)
+            @template.create-import-form.url-form(
+                result = result,
+                importFormLocalization = importFormLocalization
+            )
         </div>
         <div>
             <button
             type="submit"
-            class="text-left px-2 py-1 bg-accent text-primary rounded cursor-pointer">Submit</button>
+            class="text-left px-2 py-1 bg-accent text-primary rounded cursor-pointer">${importFormLocalization.get("submit")}</button>
         </div>
     </form>
     <script>
