+++ src/main/jte/create-import-form/content-form.jte
@@ -1,6 +1,9 @@
 @import lule.dictionary.dto.application.result.ServiceResultImp
+@import java.util.Map
 
 @param ServiceResultImp<?> result
+@param Map<String, String> importFormLocalization
+
 
 <div class="flex flex-col" id="content">
     <input type="hidden" name="url">
@@ -8,12 +11,12 @@
         <label for="area"></label>
         <textarea
         class="
-        md:min-h-200
+        md:min-h-150
         min-h-85 outline-none border border-2 border-primary rounded text-neutral resize-none
         focus:border-accent w-full"
         name="content"
         id="area"
-        placeholder="Type in content here..."></textarea>
+        placeholder="${importFormLocalization.get("space_for_content")}"></textarea>
         @if(result != null)
             @if(result.hasError() && result.messages().get("pageContent") != null)
                 <span class="text-error font-semibold text-sm">${result.messages().get("pageContent")}</span>
