+++ src/main/jte/create-import-form/url-form.jte
@@ -1,6 +1,8 @@
 @import lule.dictionary.dto.application.result.ServiceResultImp
+@import java.util.Map
 
 @param ServiceResultImp<?> result
+@param Map<String, String> importFormLocalization
 
 <div class="flex flex-col" id="content">
     <input type="hidden" name="content">
@@ -12,7 +14,7 @@
         focus:border-accent w-full"
         name="url"
         id="url"
-        placeholder="Type in URL here...">
+        placeholder="${importFormLocalization.get("space_for_url")}">
         @if(result != null)
             @if(result.hasError() && result.messages().get("url") != null)
                 <span class="text-error font-semibold text-sm">${result.messages().get("url")}</span>
