+++ src/main/resources/static/output.css
@@ -256,17 +256,8 @@
   .min-h-8 {
     min-height: calc(var(--spacing) * 8);
   }
-  .min-h-16 {
-    min-height: calc(var(--spacing) * 16);
-  }
   .min-h-18 {
     min-height: calc(var(--spacing) * 18);
-  }
-  .min-h-20 {
-    min-height: calc(var(--spacing) * 20);
-  }
-  .min-h-24 {
-    min-height: calc(var(--spacing) * 24);
   }
   .min-h-85 {
     min-height: calc(var(--spacing) * 85);
@@ -611,6 +602,11 @@
   .md\:inline {
     @media (width >= 48rem) {
       display: inline;
+    }
+  }
+  .md\:min-h-150 {
+    @media (width >= 48rem) {
+      min-height: calc(var(--spacing) * 150);
     }
   }
   .md\:min-h-200 {
